Ensure the length of a doc comment is at least 5
Previously, "/**/" was incorrectly regarded as a doc comment because it
starts with "/**" and ends with "*/". However, this caused an ICE
because some code assumed that the length of a doc comment is always 5
or longer. This commit adds an additional check to
`is_block_doc_comment` that tests the length of the input.

Fixes #28844.